requirements.txt
Django==2.0
psycopg2==2.7.3.2
gunicorn==19.7.1
Unipath==1.1
Pillow==4.3.0
django-allauth==0.34.0
django-debug-toolbar==1.9.1
